Update mena function to disregard NaNs
The previous mean extracting function has a problem with the presence of NaN's in the scan and will return NaN's which will give wrong results. So, the new version the `np.mean` function has been changed to `np.nanmean` which in case of NaN voxels will disregard these from the computations.